chore: code split UserUpdateDTO
